iteration2/requirements.txt
matplotlib>=3.8
numpy>=1.26
pandas>=2.2
PyYAML>=6.0
scikit-learn>=1.5
torch>=2.2
transformers>=4.51
